Make some changes based on slaskawi‘s comment.
