+++ core/src/main/java/org/infinispan/distexec/DefaultExecutorService.java
@@ -1100,30 +1100,4 @@ public class DefaultExecutorService extends AbstractExecutorService implements D
       }
    }
 
-   public static final class RunnableAdapter<T> implements Callable<T>, Serializable {
-
-      /** The serialVersionUID */
-      private static final long serialVersionUID = 6629286923873531028L;
-
-      protected Runnable task;
-      protected T result;
-
-      protected RunnableAdapter() {
-      }
-
-      protected RunnableAdapter(Runnable task, T result) {
-         this.task = task;
-         this.result = result;
-      }
-
-      public Runnable getTask() {
-         return task;
-      }
-
-      @Override
-      public T call() {
-         task.run();
-         return result;
-      }
-   }
 }
